refactor: move helper get-gradient-colors-for-category.helper to separate file bb-344
