+++ apps/mobile/src/libs/components/slider-content/slider-content.tsx
@@ -1,24 +1,15 @@
 import React from "react";
 
 import { GradientSlider, Text, View } from "~/libs/components/components";
-import { categoryToColors, type colorToGradientColors } from "~/libs/maps/maps";
+import { getGradientColorsForCategory } from "~/libs/helpers/helpers";
 import { globalStyles } from "~/libs/styles/styles";
 import { type SliderData } from "~/libs/types/types";
 
-import { DEFAULT_SLIDER_COLOR, MAX_SLIDER_VALUE } from "./constants/constants";
+import { MAX_SLIDER_VALUE } from "./constants/constants";
 
 type Properties = {
 	onValueChange: (id: number, sliderValue: number) => void;
 	score: SliderData;
-};
-
-const getGradientColorsForCategory = (
-	categoryName: string,
-): keyof typeof colorToGradientColors => {
-	return (
-		categoryToColors[categoryName.replaceAll(/\s+/g, "")] ??
-		DEFAULT_SLIDER_COLOR
-	);
 };
 
 const SliderContent: React.FC<Properties> = ({ onValueChange, score }) => {
