+++ apps/mobile/src/libs/helpers/helpers.ts
@@ -2,6 +2,7 @@ export { checkIfAndroid, checkIfIos } from "./check-platform/check-platform";
 export { createRadioButtonProperties } from "./create-radio-button-properties/create-radio-button-properties.helper";
 export { generateSliderGradientColors } from "./generate-slider-gradient-colors/generate-slider-gradient-colors.helper";
 export { generateSliderGradientLocations } from "./generate-slider-gradient-locations/generate-slider-gradient-locations.helper";
+export { getGradientColorsForCategory } from "./get-gradient-colors-for-category/get-gradient-colors-for-category.helper";
 export { getScreenWidth } from "./get-screen-width/get-screen-width.helper";
 export { getSecurityInputIconName } from "./get-security-input-icon-name/get-security-input-icon-name.helper";
 export { configureString, getFormattedDate } from "shared";
